add input option to specify absolute threshold for convergence checks
The absolute threshold for the residual determined in the subroutine 'cbeam3_solv_nlndyn_step' was defined as too small to detect possible very small residuals that are not detected within convergence checks by the specified relative tolerance. The absolute threshold was added to xbopts with an increased default value to detect those very small residuals and solve the issue. In addition, the user can now specify the absolute threshold which was previously hard-coded in various subroutines within xbeam.